[New feature] Generally working olivia / contestia implementation. Only decodes sometimes... Maybe needs a preamble?

+++ firmware/inc/telemetry.h
@@ -26,6 +26,6 @@
 #define TELEMETRY_H
 
 uint16_t crc_checksum(char *string);
-void timer0_tick_init(uint32_t frequency);
+void timer0_tick_init(float frequency);
 
 #endif /* TELEMETRY_H */
